+++ data/data-pipeline/pyproject.toml
@@ -25,6 +25,7 @@ ipython = "^7.31.1"
 jupyter = "^1.0.0"
 jupyter-contrib-nbextensions = "^0.5.1"
 marshmallow-dataclass = "^8.5.3"
+marshmallow-enum = "^1.5.1"
 matplotlib = "^3.4.2"
 numpy = "^1.22.1"
 pandas = "^1.2.5"
